fix: 🐛 CR fixes, tooltip firing on country bounding-box fixed
