+++ packages/widgets/src/components/choropleth/components/chart-interaction-legend/chart-interaction-legend.styles.tsx
