+++ packages/widgets/src/components/choropleth/components/chart-interaction-legend/index.ts
