+++ packages/widgets/src/components/choropleth/components/index.ts
@@ -1,4 +1,4 @@
 import GeoMatchError from './geo-match-error';
-import ChartInteractionLegend from './interaction-legend';
+import ChartInteractionLegend from './chart-interaction-legend';
 
 export { GeoMatchError, ChartInteractionLegend };
